Use risk_factors table only
