+++ backend/src/main/kotlin/fr/gouv/cnsp/monitorfish/domain/use_cases/GetVessel.kt
@@ -7,7 +7,10 @@ import fr.gouv.cnsp.monitorfish.domain.entities.VesselTrackDepth
 import fr.gouv.cnsp.monitorfish.domain.entities.VesselWithData
 import fr.gouv.cnsp.monitorfish.domain.entities.risk_factor.VesselRiskFactor
 import fr.gouv.cnsp.monitorfish.domain.exceptions.NoERSLastDepartureDateFound
-import fr.gouv.cnsp.monitorfish.domain.repositories.*
+import fr.gouv.cnsp.monitorfish.domain.repositories.ERSRepository
+import fr.gouv.cnsp.monitorfish.domain.repositories.PositionRepository
+import fr.gouv.cnsp.monitorfish.domain.repositories.RiskFactorsRepository
+import fr.gouv.cnsp.monitorfish.domain.repositories.VesselRepository
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.async
@@ -20,8 +23,6 @@ import java.time.ZonedDateTime
 class GetVessel(private val vesselRepository: VesselRepository,
                 private val positionRepository: PositionRepository,
                 private val ersRepository: ERSRepository,
-                private val currentSegmentRepository: CurrentSegmentRepository,
-                private val controlAnteriorityRepository: ControlAnteriorityRepository,
                 private val riskFactorsRepository: RiskFactorsRepository) {
     private val logger: Logger = LoggerFactory.getLogger(GetVessel::class.java)
 
@@ -72,36 +73,10 @@ class GetVessel(private val vesselRepository: VesselRepository,
             else -> ZonedDateTime.now()
         }
 
-        /*
-        TODO Si
-        gearOnboard
-        segments
-        speciesOnboard
-        controlPriorityLevel
-        segmentHighestImpact
-        segmentHighestPriority
-        numberControlsLastFiveYears
-        numberControlsLastThreeYears
-        numberInfractionsLastFiveYears
-        numberDiversionsLastFiveYears
-        numberSeizuresLastFiveYears
-        numberEscortsToQuayLastFiveYears
-        controlRateRiskFactor
-        lastControlDatetime
-        impactRiskFactor
-        probabilityRiskFactor
-        detectabilityRiskFactor
-        riskFactor
-         */
-
         return coroutineScope {
             val positionsFuture = findPositionsAsync(vesselIdentifier, internalReferenceNumber, from, to, ircs, externalReferenceNumber)
 
             val vesselFuture = async { vesselRepository.findVessel(internalReferenceNumber, externalReferenceNumber, ircs) }
-
-            val vesselCurrentSegmentFuture = async { currentSegmentRepository.findVesselCurrentSegment(internalReferenceNumber) }
-
-            val vesselControlAnteriorityFuture = async { controlAnteriorityRepository.findVesselControlAnteriority(internalReferenceNumber) }
 
             val vesselRiskFactorsFuture = async { riskFactorsRepository.findVesselRiskFactors(internalReferenceNumber) }
 
@@ -110,8 +85,6 @@ class GetVessel(private val vesselRepository: VesselRepository,
                     VesselWithData(
                             vesselFuture.await(),
                             positionsFuture.await(),
-                            vesselCurrentSegmentFuture.await(),
-                            vesselControlAnteriorityFuture.await(),
                             vesselRiskFactorsFuture.await() ?: VesselRiskFactor()
                     )
             )
