+++ backend/src/main/kotlin/fr/gouv/cnsp/monitorfish/infrastructure/database/repositories/JpaRiskFactorsRepository.kt
@@ -1,5 +1,6 @@
 package fr.gouv.cnsp.monitorfish.infrastructure.database.repositories
 
+import com.fasterxml.jackson.databind.ObjectMapper
 import fr.gouv.cnsp.monitorfish.domain.entities.risk_factor.VesselRiskFactor
 import fr.gouv.cnsp.monitorfish.domain.repositories.RiskFactorsRepository
 import fr.gouv.cnsp.monitorfish.infrastructure.database.repositories.interfaces.DBRiskFactorsRepository
@@ -10,13 +11,14 @@ import org.springframework.dao.EmptyResultDataAccessException
 import org.springframework.stereotype.Repository
 
 @Repository
-class JpaRiskFactorsRepository(private val dbRiskFactorsRepository: DBRiskFactorsRepository) : RiskFactorsRepository {
+class JpaRiskFactorsRepository(private val dbRiskFactorsRepository: DBRiskFactorsRepository,
+                               private val mapper: ObjectMapper) : RiskFactorsRepository {
     private val logger: Logger = LoggerFactory.getLogger(JpaRiskFactorsRepository::class.java)
 
     @Cacheable(value = ["risk_factors"])
     override fun findVesselRiskFactors(internalReferenceNumber: String): VesselRiskFactor? {
         try {
-            return dbRiskFactorsRepository.findByCfrEquals(internalReferenceNumber).toVesselRiskFactor()
+            return dbRiskFactorsRepository.findByCfrEquals(internalReferenceNumber).toVesselRiskFactor(mapper)
         } catch (e: EmptyResultDataAccessException) {
             logger.warn("No current segment found for CFR $internalReferenceNumber", e)
         }
