+++ frontend/cypress/integration/vessel_labels.spec.js
@@ -21,7 +21,7 @@ context('Vessel labels', () => {
     // When
     cy.get('*[data-cy^="vessel-labels"]').click({ timeout: 20000 })
 
-    cy.get('*[data-cy^="map-property-trigger"]')
+    cy.get('*[data-cy^="map-property-trigger"]', { timeout: 20000 })
       .filter(':contains("étiquettes des navires")')
       .click({ timeout: 20000 })
